requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
werkzeug==3.0.1
python-magic==0.4.27
pytesseract==0.3.10
numpy==1.26.4
Pillow==10.2.0
PyPDF2==3.0.1
transformers==4.38.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
asgiref==3.7.2
torch==2.2.1